invert pixel clock if running on a new controller
This is needed as a workaround for a pixel clock glitch in
the "new" (spring 2018) TRIK video modules.
The modules seem to have the glitch during the start of line
code but only for normal clock polarity.

Set the clock polarity in sensors (write 0x10 to reg 0x15)
AND
in the receiver via the v4l2 S_CTRL ioctl (v4l2-ctl program).