Disable UART (ttyS1) logs during Linux boot
Add a file that disables Linux boot logs on UART (ttyS1).

The console is set to an empty value using console=''. Unlike the previous version, there is no need to redirect the console to a dummy device like ttyS2, since this port does not exist on TRIK.

The u-boot-tools-native package is required to provide the mkimage utility.